+++ src/app/admin/article/page.tsx
@@ -1,0 +1,5 @@
+import { AdminArticleContent } from '@/components/admin/article/admin-article-content';
+
+export default function AdminArticlePage() {
+  return <AdminArticleContent />;
+}
