--- src/app/admin/ensiklopedia/page.tsx
@@ -1,5 +1,0 @@
-import { AdminEnsiklopediaContent } from '@/components/admin/ensiklopedia/admin-ensiklopedia-content';
-
-export default function AdminEnsiklopediaRoute() {
-  return <AdminEnsiklopediaContent />;
-}
