+++ src/app/profile/liked-article/page.tsx
@@ -1,4 +1,4 @@
-import { LikedArticlesMain } from '@/components/profile/artikel-disukai/liked-articles-main';
+import { LikedArticlesMain } from '@/components/profile/liked-article/liked-articles-main';
 import { requireUser } from '@/lib/auth/auth-page-helper';
 
 export default async function LikedArticlesPage() {
